+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Prometheus support
+- New admin bot command `admin-reset-summary` to recreate a summary comment (maintenance type command)
 
 ### Changed